update deploy predict to demo predict

+++ mindyolo/utils/metrics.py
@@ -70,7 +70,7 @@ def non_max_suppression(
         # Detections matrix nx6 (xyxy, conf, cls)
         if multi_label:
             i, j = (x[:, 5:] > conf_thres).nonzero()
-            x = np.concatenate((box[i], x[i, j + 5, None], j[:, None].astype(np.float)), 1)
+            x = np.concatenate((box[i], x[i, j + 5, None], j[:, None].astype(np.float32)), 1)
         else:  # best class only
             conf, j = x[:, 5:].max(1, keepdim=True)
             x = np.concatenate((box, conf, j.float()), 1)[conf.view(-1) > conf_thres]
